pyproject.toml
[project]
name = "smolagent-tuto"
version = "0.1.0"
description = "Agent tutorial"
email = "ruaultadrien@gmail.com"
readme = "README.md"
requires-python = ">=3.10,<3.14"

[tool.poetry]
name = "smolagent-tuto"
version = "0.1.0"
description = "Agent tutorial"
authors = ["Adrien <ruaultadrien@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
smolagents = {extras = ["litellm", "mcp"], version = "^1.18.0"}
duckduckgo-search = "^8.0.2"
gradio = "^5.31.0"
opentelemetry-sdk = "^1.33.1"
opentelemetry-exporter-otlp = "^1.33.1"
openinference-instrumentation-smolagents = "^0.1.12"
ruff = "^0.11.13"
langchain = "^0.3.25"
langchain-community = "^0.3.25"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # The following two rules are incompatible and must be configured.
    # See: https://docs.astral.sh/ruff/rules/#conflicting-rules
    "D203",
    "D212",
]